requirements.txt
Django==4.0.4
celery==5.3.1
djangorestframework==3.14.0
environs==9.5.0
GitPython==3.1.32
redis==4.6.0
requests==2.31.0
gunicorn==21.2.0
psycopg2-binary==2.9.3
gherkin-official==24.1.0
pathvalidate==3.2.0